requirements.txt
#Flask
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.13

#Flask Extensions
Flask-Classy==0.6.10

#Server
gunicorn

#Database
google-cloud==0.22.0

#Tools
python-decouple==3.0
Flask-Script==2.0.5
